Rename package to match domain name
