--- src/main/java/club/opensource/game/GamePlayer.java
@@ -1,4 +1,0 @@
-package club.opensource.game;
-
-public class GamePlayer {
-}
