--- src/main/java/club/opensource/game/Tasks.java
@@ -1,4 +1,0 @@
-package club.opensource.game;
-
-public enum Tasks {
-}
