+++ src/main/java/club/ufopensource/game/GamePlayer.java
@@ -1,0 +1,4 @@
+package club.ufopensource.game;
+
+public class GamePlayer {
+}
