+++ src/main/java/club/ufopensource/game/Tasks.java
@@ -1,0 +1,4 @@
+package club.ufopensource.game;
+
+public enum Tasks {
+}
