+++ src/main/resources/maps.yml
@@ -2,5 +2,5 @@
 # some_map:
 #   spawn: world, x, y, z
 #   tasks:
-#     - 'task_name, world, x, y, z' # list of tasks can be found in club.opensource.game.Tasks: this will map a task to a block
+#     - 'task_name, world, x, y, z' # list of tasks can be found in club.ufopensource.game.Tasks: this will map a task to a block
 #
